% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.r
\name{ENaddcontrol}
\alias{ENaddcontrol}
\title{Add a new simple control}
\usage{
ENaddcontrol(type, linkIndex, setting, nodeIndex, level)
}
\arguments{
\item{type}{the type of control to add (see details)}

\item{linkIndex}{the index of a link to control (starting from 1)}

\item{setting}{control setting applied to the link}

\item{nodeIndex}{index of the node used to control the link
(0 for EN_TIMER and EN_TIMEOFDAY controls).}

\item{level}{action level (tank level, junction pressure, or time in seconds)
that triggers the control.}
}
\value{
index index of the new control.
}
\description{
Add a new simple control
}
