% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENaddcurve}
\alias{ENaddcurve}
\title{Adds a new data curve to a project.}
\usage{
ENaddcurve(id)
}
\arguments{
\item{id}{The ID name of the curve to be added.}
}
\value{
null invisibly
}
\description{
Adds a new data curve to a project.
}
\details{
The new curve contains a single data point (1.0, 1.0).
}
