% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENadddemand}
\alias{ENadddemand}
\title{Appends a new demand to a junction node demands list.}
\usage{
ENadddemand(nodeindex, base_demand, demand_pattern, demand_name)
}
\arguments{
\item{nodeindex}{the index of a node (starting from 1).}

\item{base_demand}{the demand's base value.}

\item{demand_pattern}{the name of a time pattern used by the demand}

\item{demand_name}{the name of the demand's category}
}
\description{
Appends a new demand to a junction node demands list.
}
