% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENaddlink}
\alias{ENaddlink}
\title{Add a link to the network}
\usage{
ENaddlink(id, type, from_node, to_node)
}
\arguments{
\item{id}{name of new link}

\item{type}{of new link, see details}

\item{from_node}{id of source node for this link}

\item{to_node}{id of target node for this link}
}
\value{
index of new link
}
\description{
Add a link to the network
}
\details{
A new pipe is assigned a diameter of 10 inches (254 mm) and a length of 330
feet (~ 100 meters). Its roughness coefficient depends on the head loss formula in effect as follows:
- Hazen-Williams formula: 130
- Darcy-Weisbach formula: 0.5 millifeet (0.15 mm)
- Chezy-Manning formula: 0.01

All other pipe properties are set to 0.

A new pump has a status of EN_OPEN, a speed setting of 1, and has no pump
curve or power rating assigned to it.

A new valve has a diameter of 10 inches (254 mm) and all other properties set to 0.

Type must be one of: EN_CVPIPE, EN_PIPE, EN_PUMP, EN_PRV, EN_PSV, EN_PBV, EN_FCV, EN_TCV, EN_GPV
}
