% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENaddpattern}
\alias{ENaddpattern}
\title{Add a new time pattern}
\usage{
ENaddpattern(patternid)
}
\arguments{
\item{patternid}{the ID name of the pattern to add.}
}
\value{
invisible NULL
}
\description{
Add a new time pattern
}
\details{
The new pattern contains a single time period whose factor is 1.0.
}
