% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENaddrule}
\alias{ENaddrule}
\title{Adds a new rule-based control to a project}
\usage{
ENaddrule(rule)
}
\arguments{
\item{rule}{text of the rule following the format used in an EPANET input file.}
}
\value{
null
}
\description{
Adds a new rule-based control to a project
}
