% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENclearreport}
\alias{ENclearreport}
\title{Clears the contents of a project's report file.}
\usage{
ENclearreport()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Clears the contents of a project's report file.
}
