% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtoolkit.R
\name{ENclose}
\alias{ENclose}
\title{Close down the EPANET Toolkit system.}
\usage{
ENclose()
}
\description{
\code{ENclose} closes the EPANET Toolkit system (including all files being processed).
}
\note{
\code{ENclose} must be called when all processing has been completed, even if an error
  condition was encountered.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENclose()
}
\seealso{
\code{\link{ENopen}}
\url{http://wateranalytics.org/EPANET/group___file_management.html}
}
