% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENcloseH}
\alias{ENcloseH}
\title{close hydraulics engine}
\usage{
ENcloseH()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
\code{ENcloseH} closes the hydraulic analysis system, freeing all 
  allocated memory
}
\details{
Call \code{ENcloseH} after all hydraulics analyses have been made using 
  \code{ENinitH-ENrunH-ENnextH}. Do not call this function if \code{ENsolveH} is being used.
}
\seealso{
\code{ENopenH}, \code{ENinitH}, \code{ENrunH}, \code{ENnextH}
}
