% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENcopyreport}
\alias{ENcopyreport}
\title{Copies the current contents of a project's report file to another file}
\usage{
ENcopyreport(rptFile)
}
\arguments{
\item{rptFile}{destination file}
}
\value{
Returns NULL invisibly; called for side effect

Returns NULL invisibly; called for side effect
}
\description{
Copies the current contents of a project's report file to another file
}
\details{
This function allows toolkit clients to retrieve the contents of a project's
report file while the project is still open.
}
