% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.r
\name{ENdeletecontrol}
\alias{ENdeletecontrol}
\title{Deletes an existing simple control}
\usage{
ENdeletecontrol(controlIndex)
}
\arguments{
\item{controlIndex}{the index of the control to delete (starting from 1).}
}
\value{
null invisibly
}
\description{
Deletes an existing simple control
}
