% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENdeletecurve}
\alias{ENdeletecurve}
\title{Deletes a data curve from a project}
\usage{
ENdeletecurve(index)
}
\arguments{
\item{index}{the data curve's index (starting from 1).}
}
\value{
null invisibly
}
\description{
Deletes a data curve from a project
}
