% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENdeletedemand}
\alias{ENdeletedemand}
\title{Delete a demand from a junction node}
\usage{
ENdeletedemand(nodeindex, demandindex)
}
\arguments{
\item{nodeindex}{the index of a node (starting from 1).}

\item{demandindex}{the position of the demand in the node's demands list (starting from 1).}
}
\description{
Delete a demand from a junction node
}
