% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENdeletelink}
\alias{ENdeletelink}
\title{Delete a link from the project.}
\usage{
ENdeletelink(index, action = "EN_UNCONDITIONAL")
}
\arguments{
\item{index}{the index of the link to be deleted.}

\item{action}{The action taken if any control contains the link.}
}
\description{
Delete a link from the project.
}
\details{
If actionCode is EN_UNCONDITIONAL then the link and all simple and rule-based
controls that contain it are deleted. If set to EN_CONDITIONAL then the link
is not deleted if it appears in any control and error 261 is returned.
}
