% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENdeletenode}
\alias{ENdeletenode}
\title{Deletes a node}
\usage{
ENdeletenode(nodeindex, actionCode)
}
\arguments{
\item{nodeindex}{the index of the node to be deleted.}

\item{actionCode}{the action taken if any control contains the node and its links: EN_UNCONDITIONAL or EN_CONDITIONAL.}
}
\description{
Deletes a node
}
\details{
If `actionCode` is EN_UNCONDITIONAL then the node, its incident links and all
simple and rule-based controls that contain them are deleted. If set to
EN_CONDITIONAL then the node is not deleted if it or its incident links appear
in any controls and error code 261 is returned.
}
