% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENdeletepattern}
\alias{ENdeletepattern}
\title{Delete a new time pattern}
\usage{
ENdeletepattern(index)
}
\arguments{
\item{index}{of the pattern to delete}
}
\value{
invisible NULL
}
\description{
Delete a new time pattern
}
