% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENdeleterule}
\alias{ENdeleterule}
\title{Deletes an existing rule-based control}
\usage{
ENdeleterule(index)
}
\arguments{
\item{index}{the index of the rule to be deleted (starting from 1).}
}
\value{
null
}
\description{
Deletes an existing rule-based control
}
