% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.r
\name{ENepanet}
\alias{ENepanet}
\title{ENepanet}
\usage{
ENepanet(inpFile, rptFile, binOutFile = "")
}
\arguments{
\item{inpFile}{name of input file}

\item{rptFile}{name of report file (to be created)}

\item{binOutFile}{name of optional binary output file}
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
runs a complete EPANET simulation
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp") 
print(inp)
ENepanet( inp, "Net1.rpt") 
# try opening Net1.rpt in a text editor or reading it back
# into R with the read.rpt() function in package epanetReader
myRpt <- epanetReader::read.rpt("Net1.rpt")
summary(myRpt)
# clean-up the created file 
file.remove("Net1.rpt")
}
