% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENgetaveragepatternvalue}
\alias{ENgetaveragepatternvalue}
\title{Get average of all time factors in a pattern}
\usage{
ENgetaveragepatternvalue(index)
}
\arguments{
\item{index}{a time pattern index (starting from 1).}
}
\value{
the average value
}
\description{
Get average of all time factors in a pattern
}
