% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENgetbasedemand}
\alias{ENgetbasedemand}
\title{Gets the base demand for one of a node's demand categories.}
\usage{
ENgetbasedemand(nodeindex, demand_index = 1)
}
\arguments{
\item{nodeindex}{a node's index (starting from 1).}

\item{demand_index}{the index of a demand category for the node (starting from 1).}
}
\value{
the category's base demand.
}
\description{
Gets the base demand for one of a node's demand categories.
}
