% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{ENgetcoord}
\alias{ENgetcoord}
\title{Get coordinates for a node}
\usage{
ENgetcoord(nodeindex)
}
\arguments{
\item{index}{of node}
}
\value{
vector of x,y coordinate
}
\description{
Get coordinates for a node
}
