% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENgetcurveid}
\alias{ENgetcurveid}
\title{Retrieves the ID name of a curve given its index.}
\usage{
ENgetcurveid(index)
}
\arguments{
\item{index}{a curve's index (starting from 1).}
}
\value{
the curve's ID name.
}
\description{
Retrieves the ID name of a curve given its index.
}
