% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENgetcurveindex}
\alias{ENgetcurveindex}
\title{Retrieves the index of a curve given its ID name.}
\usage{
ENgetcurveindex(id)
}
\arguments{
\item{id}{the ID name of a curve.}
}
\value{
The curve's index (starting from 1).
}
\description{
Retrieves the index of a curve given its ID name.
}
