% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENgetcurvelen}
\alias{ENgetcurvelen}
\title{Retrieves the number of points in a curve.}
\usage{
ENgetcurvelen(index)
}
\arguments{
\item{index}{a curve's index (starting from 1).}
}
\value{
The number of data points assigned to the curve.
}
\description{
Retrieves the number of points in a curve.
}
