% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENgetcurvetype}
\alias{ENgetcurvetype}
\title{Retrieves a curve's type}
\usage{
ENgetcurvetype(index)
}
\arguments{
\item{index}{a curve's index (starting from 1).}
}
\value{
the curve's type
}
\description{
Retrieves a curve's type
}
