% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENgetcurvevalue}
\alias{ENgetcurvevalue}
\title{Retrieves the value of a single data point for a curve.}
\usage{
ENgetcurvevalue(curveIndex, pointIndex)
}
\arguments{
\item{curveIndex}{a curve's index (starting from 1).}

\item{pointIndex}{the index of a point on the curve (starting from 1).}
}
\value{
list with the point's x-value and y-value
}
\description{
Retrieves the value of a single data point for a curve.
}
