% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENgetdemandindex}
\alias{ENgetdemandindex}
\title{Retrieves the index of a node's named demand category}
\usage{
ENgetdemandindex(nodeindex, demand_name)
}
\arguments{
\item{nodeindex}{the index of a node (starting from 1).}

\item{demand_name}{the name of the demand's category}
}
\value{
demand category index
}
\description{
Retrieves the index of a node's named demand category
}
