% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENgetdemandmodel}
\alias{ENgetdemandmodel}
\title{Get type of demand model in use and its parameters}
\usage{
ENgetdemandmodel()
}
\value{
named list with parameters of the demand model
}
\description{
Get type of demand model in use and its parameters
}
