% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENgetdemandname}
\alias{ENgetdemandname}
\title{Retrieves the name of a node's demand category.}
\usage{
ENgetdemandname(nodeindex, demand_index = 1)
}
\arguments{
\item{nodeindex}{a node's index (starting from 1).}

\item{demand_index}{the index of one of the node's demand categories (starting from 1).}
}
\value{
The name of the selected category.
}
\description{
Retrieves the name of a node's demand category.
}
