% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENgetelseaction}
\alias{ENgetelseaction}
\title{Gets the properties of an ELSE action in a rule-based control.}
\usage{
ENgetelseaction(ruleIndex, actionIndex)
}
\description{
@param ruleIndex the rule's index (starting from 1).
 @param actionIndex the index of the ELSE action to retrieve (starting from 1).
 @return list with the following components:
  * linkIndex -  the index of the link in the action
  * status  - the status assigned to the link
  * setting - the value assigned to the link's setting.
}
