% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENgeterror}
\alias{ENgeterror}
\title{Returns the text of an error message generated by an error code}
\usage{
ENgeterror(errcode)
}
\arguments{
\item{errcode}{an error code.}
}
\value{
error message
}
\description{
Returns the text of an error message generated by an error code
}
