% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumps.r
\name{ENgetheadcurveindex}
\alias{ENgetheadcurveindex}
\title{Retrieves index of head curve used by a pump}
\usage{
ENgetheadcurveindex(linkindex)
}
\arguments{
\item{linkindex}{index of the pump}
}
\value{
index of head curve
}
\description{
Retrieves index of head curve used by a pump
}
