% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{ENgetlinkid}
\alias{ENgetlinkid}
\title{Retrieve the ID label of a link}
\usage{
ENgetlinkid(linkindex)
}
\arguments{
\item{linkindex}{integer specifying the link index.}
}
\value{
character ID
}
\description{
\code{ENgetlinkid} retrieves the ID label of the link given its index.
}
\note{
Link indexes are consecutive integers starting from 1.
}
\examples{
\dontrun{
  # get the ID of the 3rd link
  linkid <- ENgetlinkid(3)
}
}
\seealso{
\code{\link{ENgetlinkindex} }
}
