% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{ENgetlinkindex}
\alias{ENgetlinkindex}
\title{Retrieve the index of a link}
\usage{
ENgetlinkindex(linkid)
}
\arguments{
\item{linkid}{character}
}
\value{
integer index of requested link
}
\description{
\code{ENgetlinkindex} retrieves the index of a link with specified ID.
}
\note{
Link indexes are consecutive integers starting from 1.
}
\seealso{
\code{\link{ENgetlinkid} }
}
