% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENgetnodeid}
\alias{ENgetnodeid}
\title{Retrieve the ID label a node.}
\usage{
ENgetnodeid(nodeindex)
}
\arguments{
\item{nodeindex}{An integer node index}
}
\value{
A character string, the ID label of the specified node.
}
\description{
\code{ENgetnodeid} retrieves the ID label a node from its index
}
\note{
Node indexes are consecutive integers starting from 1.
}
\seealso{
\code{ENgetnodeindex}
\url{http://wateranalytics.org/EPANET/group___network_info.html}
}
