% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENgetnodeindex}
\alias{ENgetnodeindex}
\title{Retrieve the index of a node}
\usage{
ENgetnodeindex(nodeid)
}
\arguments{
\item{nodeid}{A character string specifying the node ID.}
}
\value{
An integer index of the specified node.
}
\description{
Retrieve the index of a node
}
\note{
Node indexes are consecutive integers starting from 1.
}
\seealso{
\code{ENgetnodeid} 
\url{http://wateranalytics.org/EPANET/group___network_info.html}
}
