% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENgetnumdemands}
\alias{ENgetnumdemands}
\title{Get number of demands for a junction node}
\usage{
ENgetnumdemands(nodeindex)
}
\arguments{
\item{nodeindex}{the index of a node (starting from 1).}
}
\value{
number of demands
}
\description{
Get number of demands for a junction node
}
