% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENgetpatternid}
\alias{ENgetpatternid}
\title{Retrieve the ID label a time pattern}
\usage{
ENgetpatternid(patternindex)
}
\arguments{
\item{patternindex}{An integer specifying the time pattern index.}
}
\value{
A character string, the pattern ID label of the specified time pattern.
}
\description{
\code{ENgetpatternid} retrieves the ID label of a particular time pattern.
}
\note{
Pattern indexes are consecutive integers starting from 1.
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___patterns.html}
}
