% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENgetpatternindex}
\alias{ENgetpatternindex}
\title{Retrieve the index a time pattern.}
\usage{
ENgetpatternindex(patternid)
}
\arguments{
\item{patternid}{A character string specifying the pattern ID}
}
\value{
An integer, the index of the specified time pattern.
}
\description{
\code{ENgetpatternindex} retrieves the index of a time pattern.
}
\note{
Pattern indexes are consecutive integers starting from 1.
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___patterns.html}
}
