% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENgetpatternlen}
\alias{ENgetpatternlen}
\title{Retrieve the number of time periods in a time pattern.}
\usage{
ENgetpatternlen(patternindex)
}
\arguments{
\item{patternindex}{An integer specifying a time pattern index.}
}
\value{
An integer, the time pattern length.
}
\description{
\code{ENgetpatternlen} retrieves the number of time periods in a specific time pattern.
}
\note{
Pattern indexes are consecutive integers starting from 1.
}
