% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENgetpremise}
\alias{ENgetpremise}
\title{Gets the properties of a premise in a rule-based control.}
\usage{
ENgetpremise(ruleIndex, premiseIndex)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{premiseIndex}{the position of the premise in the rule's list of premises  (starting from 1).}
}
\value{
list with components:
   logop the premise's logical operator ( \b IF = 1, \b AND = 2, \b OR = 3 );
    object the type of object the premise refers to (see @ref EN_RuleObject);
    objIndex the index of the object (e.g. the index of a tank);
    variable the object's variable being compared (see @ref EN_RuleVariable);
    relop the premise's comparison operator (see @ref EN_RuleOperator);
    status the status that the object's status is compared to; 
    value the value that the object's variable is compared to.
}
\description{
Gets the properties of a premise in a rule-based control.
}
