% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumps.r
\name{ENgetpumptype}
\alias{ENgetpumptype}
\title{Retrieves type of head curve used by a pump}
\usage{
ENgetpumptype(linkindex)
}
\arguments{
\item{linkindex}{index of the pump}
}
\value{
type of head curve
}
\description{
Retrieves type of head curve used by a pump
}
