% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENgetqualinfo}
\alias{ENgetqualinfo}
\title{Get quality analysis information}
\usage{
ENgetqualinfo()
}
\value{
list with elements: qualcode,  chemname, 
chemunits, tracenode
}
\description{
Get quality analysis information
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
