% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENgetresultindex}
\alias{ENgetresultindex}
\title{Retrieves the order in which a node or link appears in an output file.}
\usage{
ENgetresultindex(type, index)
}
\arguments{
\item{type}{a type of element (either EN_NODE or EN_LINK).}

\item{index}{the element's current index (starting from 1).}
}
\value{
the order in which the element's results were written to file.
}
\description{
Retrieves the order in which a node or link appears in an output file.
}
\details{
If the element does not appear in the file then its result index is 0.

 This function can be used to correctly retrieve results from an EPANET binary output file
 after the order of nodes or links in a network's database has been changed due to editing
 operations.
}
