% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENgetrule}
\alias{ENgetrule}
\title{Retrieves summary information about a rule-based control.}
\usage{
ENgetrule(index)
}
\arguments{
\item{index}{the rule's index (starting from 1).}
}
\value{
list with components:
 nPremises number of premises in the rule's IF section;
 nThenActions number of actions in the rule's THEN section;
 nElseActions number of actions in the rule's ELSE section;
 priority the rule's priority value.
}
\description{
Retrieves summary information about a rule-based control.
}
