% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENgetruleID}
\alias{ENgetruleID}
\title{Gets the ID name of a rule-based control given its index.}
\usage{
ENgetruleID(index)
}
\arguments{
\item{index}{the rule's index (starting from 1).}
}
\value{
rule's ID name.
}
\description{
Gets the ID name of a rule-based control given its index.
}
