% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENgetstatistic}
\alias{ENgetstatistic}
\title{Analysis convergence statistics.}
\usage{
ENgetstatistic(stat)
}
\arguments{
\item{stat}{one of the statistics tabulated below}
}
\value{
value of the stat
}
\description{
Analysis convergence statistics.
}
\details{
These statistics report the convergence criteria for the most current
hydraulic analysis and the cumulative water quality mass balance error at the
current simulation time. 

\tabular{ll}{
  \code{EN_ITERATIONS} \tab Number of hydraulic iterations taken. \cr
  \code{EN_RELATIVEERROR} \tab Sum of link flow changes / sum of link flows. \cr
  \code{EN_MAXHEADERROR} \tab Largest head loss error for links. \cr
  \code{EN_MAXFLOWCHANGE} \tab Largest flow change in links. \cr
  \code{EN_MASSBALANCE} \tab Cumulative water quality mass balance ratio. \cr
  \code{EN_DEFICIENTNODES} \tab Number of pressure deficient nodes. \cr
  \code{EN_DEMANDREDUCTION} \tab % demand reduction at pressure deficient nodes \cr
}
}
