% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.r
\name{ENgetversion}
\alias{ENgetversion}
\title{Retrieve the current version number of the EPANET Toolkit.}
\usage{
ENgetversion()
}
\value{
An integer, the Toolkit version number.
}
\description{
\code{ENgetversion} retrieves the current version number of the EPANET Toolkit.
}
\note{
The version number is a 5-digit integer that increases sequentially from
  20001 with each new update of the Toolkit.
}
\examples{
ENgetversion()
}
