% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENgetvertex}
\alias{ENgetvertex}
\title{Get vertex coordinates}
\usage{
ENgetvertex(index, vertex)
}
\arguments{
\item{index}{a link's index (starting from 1).}

\item{vertex}{index of vertex for getting coords}
}
\value{
list with elements x and y
}
\description{
Get vertex coordinates
}
