% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENgetvertexcount}
\alias{ENgetvertexcount}
\title{Get number of vertices for a link}
\usage{
ENgetvertexcount(index)
}
\arguments{
\item{index}{a link's index (starting from 1).}
}
\value{
number of indices
}
\description{
Get number of vertices for a link
}
