% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.r
\name{ENinit}
\alias{ENinit}
\title{ENinit}
\usage{
ENinit(rptFile, outFile, unitsType, headLossType)
}
\arguments{
\item{rptFile}{the name of a report file to be created (or "" if not needed).}

\item{outFile}{the name of a binary output file to be created (or "" if not needed).}

\item{unitsType}{the choice of flow units. One of: "EN_CFS", "EN_GPM", "EN_MGD",
"EN_IMGD", "EN_AFD", "EN_LPS", "EN_LPM", "EN_MLD", "EN_CMH", "EN_CMD"}

\item{headLossType}{the choice of head loss formula . One of: EN_HW, EN_DW, EN_CM}
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Initializes an empty EPANET network
}
\details{
This function should be called to create an empty EPANET project without 
an EPANET-formatted input file.  If the
project receives it's network data from an input file then there is no need to
call this function; use ENopen instead.
}
