% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENnextQ}
\alias{ENnextQ}
\title{Advances WQ simulation to start of the next hydraulic time period.}
\usage{
ENnextQ()
}
\value{
seconds until next hydraulic event occurs or
         0 if at the end of the simulation period.
}
\description{
Advances WQ simulation to start of the next hydraulic time period.
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
