% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENopenH}
\alias{ENopenH}
\title{Open hydraulics analysis system.}
\usage{
ENopenH()
}
\description{
\code{ENopenH} opens the EPANET hydraulics analysis system.
}
\details{
Call \code{ENopenH} prior to running the first hydraulic analysis using the 
  \code{ENinitH-ENrunH-ENnextH} sequence. Multiple analyses can be made before calling 
  \code{ENcloseH} to close the hydraulic analysis system.
  
  Do not call this function if \code{ENsolveH} is being used to run a complete hydraulic analysis.
}
\seealso{
\code{ENinitH}, \code{ENrunH}, \code{ENnextH}, \code{ENcloseH}
}
