% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENopenQ}
\alias{ENopenQ}
\title{Sets up for Water Quality analysis}
\usage{
ENopenQ()
}
\value{
Returns NULL invisibly on success or throws an error or warning
}
\description{
Sets up for Water Quality analysis
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
