% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENresetreport}
\alias{ENresetreport}
\title{Resets a project's report options to their default values.}
\usage{
ENresetreport()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Resets a project's report options to their default values.
}
\details{
After calling this function the default reporting options are in effect. These are:
\itemize{
\item no status report
\item no energy report
\item no nodes reported on
\item no links reported on
\item node variables reported to 2 decimal places
\item link variables reported to 2 decimal places (3 for friction factor)
\item node variables reported are elevation, head, pressure, and quality
\item link variables reported are flow, velocity, and head loss.
}
}
