% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENrunQ}
\alias{ENrunQ}
\title{Computs WQ results at current time .}
\usage{
ENrunQ()
}
\value{
current simulation time in seconds
}
\description{
Computs WQ results at current time .
}
\details{
used in a loop with ENnextQ() to run
an extended period WQ simulation.
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
