% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENsaveH}
\alias{ENsaveH}
\title{ENsaveH
 
Saves hydraulic results to binary file}
\usage{
ENsaveH()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
ENsaveH
 
Saves hydraulic results to binary file
}
\details{
Must be called before ENreport() if no WQ simulation has been made.
Should not be called if ENsolveQ() will be used.
}
