% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENsavehydfile}
\alias{ENsavehydfile}
\title{Saves temporary hydraulics file to disk}
\usage{
ENsavehydfile(hydfile)
}
\arguments{
\item{hydfile}{the name of the file to be created.}
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Saves temporary hydraulics file to disk
}
\details{
Use this function to save the current set of hydraulics results to a file, either for
 post-processing or to be used at a later time by calling \code{ENusehydfile}.

 The hydraulics file contains nodal demands and heads and link flows, status, and settings
 for all hydraulic time steps, even intermediate ones.

 Before calling this function hydraulic results must have been generated and saved by having
 called \code{ENsolveH} or the \code{ENinitH} - \code{ENrunH} - \code{ENnextH} sequence with the initflag
 argument of \code{ENinitH} set to EN_SAVE or EN_SAVE_AND_INIT.
}
