% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet.r
\name{ENsaveinpfile}
\alias{ENsaveinpfile}
\title{ENaveinpfile
  
Saves current data to "INP" formatted text file.}
\usage{
ENsaveinpfile(filename)
}
\arguments{
\item{filename}{The file path to create}
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
ENaveinpfile
  
Saves current data to "INP" formatted text file.
}
