% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENsetbasedemand}
\alias{ENsetbasedemand}
\title{Sets the base demand for one of a node's demand categories.}
\usage{
ENsetbasedemand(nodeindex, demand_index = 1, base_demand)
}
\arguments{
\item{nodeindex}{a node's index (starting from 1).}

\item{demand_index}{the index of a demand category for the node (starting from 1).}

\item{base_demand}{the category's base demand.}
}
\description{
Sets the base demand for one of a node's demand categories.
}
