% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{ENsetcoord}
\alias{ENsetcoord}
\title{Set coordinates for a node}
\usage{
ENsetcoord(nodeindex, x, y)
}
\arguments{
\item{nodeindex}{index of nodes for which to set coords}

\item{x}{coordinate}

\item{y}{coordinate}
}
\value{
returns NULL invisibily on success or raises an error or warning
}
\description{
Set coordinates for a node
}
