% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENsetcurveid}
\alias{ENsetcurveid}
\title{Changes the ID name of a data curve given its index.}
\usage{
ENsetcurveid(index, id)
}
\arguments{
\item{index}{a data curve index (starting from 1).}

\item{id}{the data curve's new ID name.}
}
\value{
null invisibly
}
\description{
Changes the ID name of a data curve given its index.
}
