% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.r
\name{ENsetcurvevalue}
\alias{ENsetcurvevalue}
\title{Sets the value of a single data point for a curve.}
\usage{
ENsetcurvevalue(curveIndex, pointIndex, x, y)
}
\arguments{
\item{curveIndex}{a curve's index (starting from 1).}

\item{pointIndex}{the index of a point on the curve (starting from 1).}

\item{x}{the point's new x-value.}

\item{y}{the point's new y-value.}
}
\value{
null invisibly
}
\description{
Sets the value of a single data point for a curve.
}
