% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENsetdemandname}
\alias{ENsetdemandname}
\title{Sets the name of a node's demand category.}
\usage{
ENsetdemandname(nodeindex, demand_index, demand_name)
}
\arguments{
\item{nodeindex}{a node's index (starting from 1).}

\item{demand_index}{the index of one of the node's demand categories (starting from 1).}

\item{demand_name}{The name of the selected category. No more than 30 characters}
}
\description{
Sets the name of a node's demand category.
}
