% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.r
\name{ENsetdemandpattern}
\alias{ENsetdemandpattern}
\title{Sets the index of a time pattern used for one of a node's demand categories.}
\usage{
ENsetdemandpattern(nodeindex, demand_index, pattern_index)
}
\arguments{
\item{nodeindex}{a node's index (starting from 1).}

\item{demand_index}{the index of one of the node's demand categories (starting from 1).}

\item{pattern_index}{the index of the time pattern assigned to the category.}
}
\description{
Sets the index of a time pattern used for one of a node's demand categories.
}
\details{
Specifying a pattern index of 0 indicates that no time pattern is assigned to the 
demand category.
}
