% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetelseaction}
\alias{ENsetelseaction}
\title{Set properties of an ELSE action in a rule-based control}
\usage{
ENsetelseaction(ruleIndex, actionIndex, linkIndex, status, setting)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{actionIndex}{the index of the ELSE action being modified (starting from 1).}

\item{linkIndex}{the index of the link in the action (starting from 1).}

\item{status}{the new status assigned to the link}

\item{setting}{the new value assigned to the link's setting.}
}
\value{
null invisibly
}
\description{
Set properties of an ELSE action in a rule-based control
}
