% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\name{ENsetflowunits}
\alias{ENsetflowunits}
\title{Sets flow units.}
\usage{
ENsetflowunits(units)
}
\arguments{
\item{units}{the choice of flow units. One of: "EN_CFS", "EN_GPM", "EN_MGD",
"EN_IMGD", "EN_AFD", "EN_LPS", "EN_LPM", "EN_MLD", "EN_CMH", "EN_CMD"}
}
\value{
null invisibly
}
\description{
Sets flow units.
}
\details{
Flow units in liters or cubic meters implies that SI metric units
are used for all other quantities in addition to flow. Otherwise US Customary
units are employed.
}
