% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumps.r
\name{ENsetheadcurveindex}
\alias{ENsetheadcurveindex}
\title{Sets index of head curve used by a pump}
\usage{
ENsetheadcurveindex(linkindex, curveindex)
}
\arguments{
\item{linkindex}{index of the pump}

\item{curveindex}{index of head curve to assign}
}
\value{
null invisibly
}
\description{
Sets index of head curve used by a pump
}
