% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENsetlinkid}
\alias{ENsetlinkid}
\title{Change the ID of a link}
\usage{
ENsetlinkid(index, newid)
}
\arguments{
\item{index}{of the target link}

\item{newid}{new name for the link (no more than 30 characters)}
}
\description{
Change the ID of a link
}
