% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENsetlinknodes}
\alias{ENsetlinknodes}
\title{Set the indexes of a link's start- and end-nodes}
\usage{
ENsetlinknodes(index, node1_index, node2_index)
}
\arguments{
\item{index}{a link's index (starting from 1).}

\item{node1_index}{The index of the link's start node (starting from 1).}

\item{node2_index}{The index of the link's end node (starting from 1).}
}
\description{
Set the indexes of a link's start- and end-nodes
}
