% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENsetlinktype}
\alias{ENsetlinktype}
\title{Change a link's type}
\usage{
ENsetlinktype(index, type, action = "EN_UNCONDITIONAL")
}
\arguments{
\item{index}{of link before type change}

\item{type}{the new type to change the link to (see details)}

\item{action}{the action taken if any controls contain the link (see details)}
}
\value{
link index after the type change
}
\description{
Change a link's type
}
\details{
Link type is one of: EN_CVPIPE, EN_PIPE, EN_PUMP, EN_PRV, EN_PSV, EN_PBV, EN_FCV, EN_TCV, EN_GPV

If actionCode is EN_UNCONDITIONAL then all simple and rule-based controls that
contain the link are deleted when the link's type is changed. If set to
EN_CONDITIONAL then the type change is cancelled if the link appears in any
control and error 261 is returned.
}
