% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENsetnodeid}
\alias{ENsetnodeid}
\title{Changes the ID name of a node}
\usage{
ENsetnodeid(nodeindex, newid)
}
\arguments{
\item{nodeindex}{index of the node}

\item{newid}{new ID name of the node}
}
\description{
Changes the ID name of a node
}
