% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENsetpatternid}
\alias{ENsetpatternid}
\title{Change the ID name of a time pattern given its index.}
\usage{
ENsetpatternid(index, id)
}
\arguments{
\item{index}{a time pattern index (starting from 1).}

\item{id}{the time pattern's new ID name.}
}
\value{
NULL invisibly
}
\description{
Change the ID name of a time pattern given its index.
}
