% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetpremise}
\alias{ENsetpremise}
\title{Sets the properties of a premise in a rule-based control.}
\usage{
ENsetpremise(
  ruleIndex,
  premiseIndex,
  logop,
  object,
  objIndex,
  variable,
  relop,
  status,
  value
)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{premiseIndex}{the position of the premise in the rule's list of premises.}

\item{logop}{the premise's logical operator ( \b IF = 1, \b AND = 2, \b OR = 3 ).}

\item{object}{the type of object the premise refers to (see @ref EN_RuleObject).}

\item{objIndex}{the index of the object (e.g. the index of a tank).}

\item{variable}{the object's variable being compared (see @ref EN_RuleVariable).}

\item{relop}{the premise's comparison operator (see @ref EN_RuleOperator).}

\item{status}{the status that the object's status is compared to (see @ref EN_RuleStatus).}

\item{value}{the value that the object's variable is compared to.}
}
\value{
null
}
\description{
Sets the properties of a premise in a rule-based control.
}
