% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetpremiseindex}
\alias{ENsetpremiseindex}
\title{Sets the index of an object in a premise of a rule-based control}
\usage{
ENsetpremiseindex(ruleIndex, premiseIndex, objIndex)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{premiseIndex}{the premise's index (starting from 1).}

\item{objIndex}{the index of the premise's object (e.g. the index of a tank).}
}
\value{
null
}
\description{
Sets the index of an object in a premise of a rule-based control
}
