% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetpremisestatus}
\alias{ENsetpremisestatus}
\title{Sets the status being compared to in a premise of a rule-based control}
\usage{
ENsetpremisestatus(ruleIndex, premiseIndex, status)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{premiseIndex}{the premise's index (starting from 1).}

\item{status}{the status that the premise's object status is compared to (see @ref EN_RuleStatus).}
}
\value{
null
}
\description{
Sets the status being compared to in a premise of a rule-based control
}
