% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetpremisevalue}
\alias{ENsetpremisevalue}
\title{Sets the value in a premise of a rule-based control}
\usage{
ENsetpremisevalue(ruleIndex, premiseIndex, value)
}
\arguments{
\item{ruleIndex}{the rule's index (staring from 1).}

\item{premiseIndex}{the premise's index (starting from 1).}

\item{value}{The value that the premise's variable is compared to.}
}
\value{
null
}
\description{
Sets the value in a premise of a rule-based control
}
