% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENsetreport}
\alias{ENsetreport}
\title{Processes a reporting format command.}
\usage{
ENsetreport(format)
}
\arguments{
\item{format}{report formatting command:  one line from the [REPORT] section of an inp file}
}
\description{
Processes a reporting format command.
}
