% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetrulepriority}
\alias{ENsetrulepriority}
\title{Sets the priority of a rule-based control.}
\usage{
ENsetrulepriority(index, priority)
}
\arguments{
\item{index}{the rule's index (starting from 1).}

\item{priority}{the priority value assigned to the rule.}
}
\value{
null
}
\description{
Sets the priority of a rule-based control.
}
