% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENsetstatusreport}
\alias{ENsetstatusreport}
\title{Sets the level of hydraulic status reporting.}
\usage{
ENsetstatusreport(level)
}
\arguments{
\item{level}{one of: EN_NO_REPORT, EN_NORMAL_REPORT, EN_FULL_REPORT}
}
\description{
Sets the level of hydraulic status reporting.
}
