% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENsettankdata}
\alias{ENsettankdata}
\title{Sets properties for a tank}
\usage{
ENsettankdata(
  nodeindex,
  elevation,
  init_level,
  min_level,
  max_level,
  diameter,
  min_volume,
  volume_curve = ""
)
}
\arguments{
\item{nodeindex}{tank's node index (starting from 1)}

\item{elevation}{the tank's bottom elevation.}

\item{init_level}{the initial water level in the tank.}

\item{min_level}{the minimum water level for the tank.}

\item{max_level}{the maximum water level for the tank.}

\item{diameter}{the tank's diameter (0 if a volume curve is supplied).}

\item{min_volume}{the volume of the tank at its minimum water level.}

\item{volume_curve}{the name of the tank's volume curve ("" for no curve)}
}
\description{
Sets properties for a tank
}
