% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetthenaction}
\alias{ENsetthenaction}
\title{Sets the properties of a THEN action in a rule-based control.}
\usage{
ENsetthenaction(ruleIndex, actionIndex, linkIndex, status, setting)
}
\description{
@param ruleIndex the rule's index (starting from 1).
 @param actionIndex the index of the THEN action to modify (starting from 1).
 @param linkIndex the index of the link in the action.
 @param status the new status assigned to the link (see @ref EN_RuleStatus).
 @param setting the new value assigned to the link's setting.
 @return null
}
