% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links22.r
\name{ENsetvertices}
\alias{ENsetvertices}
\title{Set a link's vertices}
\usage{
ENsetvertices(index, x, y)
}
\arguments{
\item{index}{a link's index}

\item{x}{numeric vector of x-coordinates}

\item{y}{numeric vector of y-coordinates}
}
\description{
Set a link's vertices
}
