% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENsolveH}
\alias{ENsolveH}
\title{ENsolveH}
\usage{
ENsolveH()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Solves the network hydraulics for all time periods
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt","Net1.bin")
ENsolveH()
ENsolveQ()
ENgetnodevalue(2, "EN_PRESSURE")
ENclose() 
# clean-up the created files
file.remove("Net1.rpt") 
file.remove("Net1.bin") 
}
