% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENstepQ}
\alias{ENstepQ}
\title{Advances WQ simulation one water quality time step.}
\usage{
ENstepQ()
}
\value{
time remaining in the overall simulation
}
\description{
Advances WQ simulation one water quality time step.
}
