% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.r
\name{ENusehydfile}
\alias{ENusehydfile}
\title{Uses previously saved binary hydraulics file to supply a project's hydraulics.}
\usage{
ENusehydfile(hydfile)
}
\arguments{
\item{hydfile}{name of file containing hydraulic results}
}
\description{
Uses previously saved binary hydraulics file to supply a project's hydraulics.
}
\details{
Call this function to re-use a set of hydraulic analysis results saved previously. This
 can save computational time if water quality analyses are being made under the same set
 of hydraulic conditions.

 Do not call this function while the hydraulics solver is open.
}
