% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet2toolkit.r
\docType{package}
\name{epanet2toolkit-package}
\alias{epanet2toolkit}
\alias{epanet2toolkit-package}
\title{epanet2toolkit: Call 'EPANET' Functions to Simulate Pipe Networks}
\description{
Enables simulation of water piping networks using 'EPANET'. The package provides functions from the 'EPANET' programmer's toolkit as R functions so that basic or customized simulations can be carried out from R. The package uses 'EPANET' version 2.2 from Open Water Analytics \url{https://github.com/OpenWaterAnalytics/EPANET/releases/tag/v2.2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bradleyjeck/epanet2toolkit}
}

}
\author{
\strong{Maintainer}: Bradley Eck \email{brad@bradeck.net}

Authors:
\itemize{
  \item Ernesto Arandia \email{earandia.ie@gmail.com}
  \item Lew Rossman
}

Other contributors:
\itemize{
  \item Michael Tryby [contributor]
  \item Sam Hatchett [contributor]
  \item Feng Shang [contributor]
  \item James Uber [contributor]
  \item Tom Taxon [contributor]
  \item Hyoungmin Woo [contributor]
  \item Jinduan Chen [contributor]
  \item Yunier Soad [contributor]
  \item Mike Kane [contributor]
  \item Demetrios Eliades [contributor]
  \item Will Furnass [contributor]
  \item Steffen Macke [contributor]
  \item Marios Kyriakou [contributor]
  \item Elad Salomons [contributor]
  \item Maurizio Cingi [contributor]
  \item Bryant McDonnell [contributor]
  \item Angela Marchi [contributor]
  \item Markus Sunela [contributor]
  \item Milad Ghiami [contributor]
  \item IBM Corp. [copyright holder]
  \item Open Water Analytics [copyright holder]
}

}
\keyword{internal}
