% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet2toolkit.r
\docType{package}
\name{epanet2toolkit}
\alias{epanet2toolkit}
\title{epanet2toolkit}
\description{
Package for using EPANET 2 from R.  Run a full EPANET simulation
using \code{\link{ENepanet}} or build a custom simulation starting with toolkit
functions like \code{\link{ENopen}}.
}
\author{
Ernesto Arandia & Bradley J. Eck
}
