% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{getOpenflag}
\alias{getOpenflag}
\title{getOpenflag}
\usage{
getOpenflag()
}
\value{
boolean
}
\description{
gets current value of EPANET global variable Openflag
}
\details{
primary use of this function is to allow
package functions to check status of toolkit before 
calling a function
}
\examples{
getOpenflag()
}
