% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{getOpenHflag}
\alias{getOpenHflag}
\title{getOpenHflag}
\usage{
getOpenHflag()
}
\value{
boolean
}
\description{
gets current value of EPANET global variable OpenHflag
}
\details{
primary use of this function is to allow
package functions to check open/closed status of the 
toolkit's hydraulic solver before 
calling a hydraulic function
}
\examples{
getOpenHflag()
}
