% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENcloseQ}
\alias{ENcloseQ}
\title{Close water quality analysis and free allocated memory}
\usage{
ENcloseQ()
}
\description{
Close water quality analysis and free allocated memory
}
\details{
Do not call this function if ENsolveQ is being used.
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
