% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{ENgetcoord}
\alias{ENgetcoord}
\title{Get coordinates for a node}
\usage{
ENgetcoord(nodeindex)
}
\arguments{
\item{nodeindex}{of node}
}
\value{
vector of x,y coordinate
}
\description{
Get coordinates for a node
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetcoord(3)
ENclose()
}
