% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENgetelseaction}
\alias{ENgetelseaction}
\title{Gets the properties of an ELSE action in a rule-based control.}
\usage{
ENgetelseaction(ruleIndex, actionIndex)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1).}

\item{actionIndex}{the index of the ELSE action to retrieve (starting from 1).}
}
\value{
list with the following components:
  \describe{
  \item{linkIndex}{ the index of the link in the action}
  \item{status}{the status assigned to the link}
  \item{setting}{the value assigned to the link's setting}
}
}
\description{
Gets the properties of an ELSE action in a rule-based control.
}
