% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{ENgetlinkid}
\alias{ENgetlinkid}
\title{Retrieve the ID label of a link}
\usage{
ENgetlinkid(linkindex)
}
\arguments{
\item{linkindex}{integer specifying the link index.}
}
\value{
character ID
}
\description{
\code{ENgetlinkid} retrieves the ID label of the link given its index.
}
\note{
Link indexes are consecutive integers starting from 1.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetlinkid(1)
ENgetlinkid(12)
ENclose()
}
\seealso{
\code{\link{ENgetlinkindex} }
}
