% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{ENgetlinkindex}
\alias{ENgetlinkindex}
\title{Retrieve the index of a link}
\usage{
ENgetlinkindex(linkid)
}
\arguments{
\item{linkid}{character}
}
\value{
integer index of requested link
}
\description{
\code{ENgetlinkindex} retrieves the index of a link with specified ID.
}
\note{
Link indexes are consecutive integers starting from 1.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetlinkindex("22")
ENclose()
}
\seealso{
\code{\link{ENgetlinkid} }
}
