% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENgetnodeid}
\alias{ENgetnodeid}
\title{Retrieve the ID label a node.}
\usage{
ENgetnodeid(nodeindex)
}
\arguments{
\item{nodeindex}{An integer node index}
}
\value{
A character string, the ID label of the specified node.
}
\description{
\code{ENgetnodeid} retrieves the ID label a node from its index
}
\note{
Node indexes are consecutive integers starting from 1.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetnodeid(1)
ENgetnodeid(5)
ENgetnodeid(9)
ENclose()
}
\seealso{
\code{ENgetnodeindex}
\url{http://wateranalytics.org/EPANET/group___network_info.html}
}
