% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.r
\name{ENgetnodeindex}
\alias{ENgetnodeindex}
\title{Retrieve the index of a node}
\usage{
ENgetnodeindex(nodeid)
}
\arguments{
\item{nodeid}{A character string specifying the node ID.}
}
\value{
An integer index of the specified node.
}
\description{
Retrieve the index of a node
}
\note{
Node indexes are consecutive integers starting from 1.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetnodeindex("10")
ENgetnodeindex("23")
ENclose()
}
\seealso{
\code{ENgetnodeid} 
\url{http://wateranalytics.org/EPANET/group___network_info.html}
}
