% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENgetpatternlen}
\alias{ENgetpatternlen}
\title{Retrieve the number of time periods in a time pattern.}
\usage{
ENgetpatternlen(patternindex)
}
\arguments{
\item{patternindex}{An integer specifying a time pattern index.}
}
\value{
An integer, the time pattern length.
}
\description{
\code{ENgetpatternlen} retrieves the number of time periods in a specific time pattern.
}
\note{
Pattern indexes are consecutive integers starting from 1.
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetpatternlen(1)
ENclose()
}
