% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENopenQ}
\alias{ENopenQ}
\title{Sets up for Water Quality analysis}
\usage{
ENopenQ()
}
\value{
Returns NULL invisibly on success or throws an error or warning
}
\description{
Sets up for Water Quality analysis
}
\examples{
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENsolveH()
ENsetqualtype("EN_CHEM", "Chlorine", "mg/L", "")
ENopenQ()
ENinitQ(0)
ENrunQ()
ENcloseQ()
ENclose()
# clean-up the created files
file.remove("Net1.rpt") 
}
