% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ENreport}
\alias{ENreport}
\title{Write simulation report to the report file}
\usage{
ENreport()
}
\value{
Returns NULL invisibly; called for side effect
}
\description{
Write simulation report to the report file
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt", "Net1.bin")
ENsolveH()
ENsolveQ()
ENreport()
ENclose()
# clean up the created files 
file.remove("Net1.rpt") 
file.remove("Net1.bin") 
}
