% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.r
\name{ENrunQ}
\alias{ENrunQ}
\title{Computs WQ results at current time .}
\usage{
ENrunQ()
}
\value{
current simulation time in seconds
}
\description{
Computs WQ results at current time .
}
\details{
used in a loop with ENnextQ() to run
an extended period WQ simulation.
}
\examples{
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENsolveH()
ENsetqualtype("EN_CHEM", "Chlorine", "mg/L", "")
ENopenQ()
ENinitQ(0)
ENrunQ()
ENcloseQ()
}
\seealso{
\url{http://wateranalytics.org/EPANET/group___quality_functions.html}
}
