% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{ENsetcoord}
\alias{ENsetcoord}
\title{Set coordinates for a node}
\usage{
ENsetcoord(nodeindex, x, y)
}
\arguments{
\item{nodeindex}{index of nodes for which to set coords}

\item{x}{coordinate}

\item{y}{coordinate}
}
\value{
returns NULL invisibily on success or raises an error or warning
}
\description{
Set coordinates for a node
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetcoord(3)
ENsetcoord(3,33,44)
ENgetcoord(3)
ENclose()
}
