% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.r
\name{ENsetpatternvalue}
\alias{ENsetpatternvalue}
\title{set pattern value}
\usage{
ENsetpatternvalue(index, period, value)
}
\arguments{
\item{index}{index of pattern}

\item{period}{time period for setting the value}

\item{value}{value to set}
}
\value{
returns NULL inivisbly on success
}
\description{
set pattern value
}
\examples{
# path to Net1.inp example file included with this package
inp <- file.path( find.package("epanet2toolkit"), "extdata","Net1.inp")  
ENopen( inp, "Net1.rpt")
ENgetpatternvalue(1,3)
ENsetpatternvalue(1,3, 9.876)
ENgetpatternvalue(1,3)
ENclose()
}
