% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.r
\name{ENsetthenaction}
\alias{ENsetthenaction}
\title{Set properties of THEN action in a rule-based control}
\usage{
ENsetthenaction(ruleIndex, actionIndex, linkIndex, status, setting)
}
\arguments{
\item{ruleIndex}{the rule's index (starting from 1)}

\item{actionIndex}{the index of the THEN action to modify (starting from 1)}

\item{linkIndex}{the index of the link in the action}

\item{status}{the new status assigned to the link}

\item{setting}{the new value assigned to the link's setting}
}
\value{
null
}
\description{
Set properties of THEN action in a rule-based control
}
